+++ Veronique/Veronique.Test/Commands/SimpleCommandMock.cs
@@ -19,15 +19,38 @@
 #region Namespaces
 using System.Collections.Generic;
 using System.IO;
+using Veronique.Commands;
 #endregion
 
-namespace Veronique.Commands
+namespace Veronique.Test.CommandStubs
 {
     /// <summary>
-    /// Provides the ability to execute an external command and return the output of the command.
+    /// Represents a simple mock-up for an <see cref="ICommand"/> implementation returning a static result.
     /// </summary>
-    public interface ICommand
+    public class SimpleCommandMock : ICommand
     {
+        #region Constants and Fields
+
+        /// <summary>
+        /// The result that shall be returned.
+        /// </summary>
+        private string result;
+
+        #endregion
+
+        #region Constructors and Destructors
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SimpleCommandMock"/> class.
+        /// </summary>
+        /// <param name="stubResult">The result that shall be returned.</param>
+        public SimpleCommandMock(string stubResult)
+        {
+            this.result = stubResult;
+        }
+
+        #endregion
+
         #region Methods
 
         /// <summary>
@@ -35,7 +58,10 @@ namespace Veronique.Commands
         /// </summary>
         /// <returns>The output of the external command.</returns>
         /// <exception cref="IOException">An error occured while executing the external command.</exception>
-        string Execute();
+        public string Execute()
+        {
+            return this.Execute(new List<string>());
+        }
 
         /// <summary>
         /// Executes the command and returns the output.
@@ -43,7 +69,10 @@ namespace Veronique.Commands
         /// <param name="arguments">The arguments that shall be given to the external application.</param>
         /// <returns>The output of the external command.</returns>
         /// <exception cref="IOException">An error occured while executing the external command.</exception>
-        string Execute(IEnumerable<string> arguments);
+        public string Execute(IEnumerable<string> arguments)
+        {
+            return this.result;
+        }
 
         #endregion
     }